requirements-dev.txt
ipdb
pandas
pytest
wheel
setuptools
twine
build
python-dotenv